+++ contracts/contracts-upgradable/token/OFT20/OFT20Upgradeable.sol
@@ -10,14 +10,12 @@ import "./IOFT20Upgradeable.sol";
 
 // override decimal() function is needed
 contract OFT20Upgradeable is Initializable, OFT20CoreUpgradeable, ERC20Upgradeable, IOFT20Upgradeable {
-
     function __OFT20Upgradeable_init(string memory _name, string memory _symbol, address _lzEndpoint) internal onlyInitializing {
         __ERC20_init_unchained(_name, _symbol);
         __OFT20CoreUpgradeable_init_unchained(_lzEndpoint);
     }
 
-    function __OFT20Upgradeable_init_unchained(string memory _name, string memory _symbol, address _lzEndpoint) internal onlyInitializing {
-    }
+    function __OFT20Upgradeable_init_unchained(string memory _name, string memory _symbol, address _lzEndpoint) internal onlyInitializing {}
 
     function supportsInterface(bytes4 interfaceId) public view virtual override(OFT20CoreUpgradeable, IERC165Upgradeable) returns (bool) {
         return interfaceId == type(IOFT20Upgradeable).interfaceId || interfaceId == type(IERC20Upgradeable).interfaceId || super.supportsInterface(interfaceId);
@@ -42,5 +40,5 @@ contract OFT20Upgradeable is Initializable, OFT20CoreUpgradeable, ERC20Upgradeab
      * variables without shifting down storage in the inheritance chain.
      * See https://docs.openzeppelin.com/contracts/4.x/upgradeable#storage_gaps
      */
-    uint256[50] private __gap;
+    uint[50] private __gap;
 }
